+++ packages/curvenote-cli/src/sync/init/peopleLookup.ts
